Countdown and Arcade mode
Added a countdown to prepare the user for the start of the game. I also added a mod to have play mode and arcade mode. The countdown class also deals with all countdowns but also fires up the GameManager.StartGame function which I don't think it should!! possibly add events to handle those calls